pyproject.toml
[tool.poetry]
name = "shortcut-summarizer"
version = "0.1.0"
description = ""
authors = ["RxSaml"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = { version = "^2", source="torch"}
pydantic = "^2.10.6"
requests = "^2.32.3"
pyfunctional = "^1.5.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
types-requests = "^2.32.0.20241016"
pylint = "^3.3.3"
flake8 = "^7.1.1"
pre-commit = "^4.1.0"
black = "^24.10.0"
flake8-black = "^0.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/rocm6.2"
secondary = true

[tool.poetry.plugins.dotenv]
ignore = "false"
location = ".env.dev"

[tool.black]
line-length = 79
target-version = ["py39"]
